subtypes & and deprecation to Js.t
- Js.t is not needed any more & will be removed in the future releases
