+++ jscomp/common/js_config.ml
@@ -75,8 +75,6 @@ let force_cmj = ref false
 
 let jsx_version = ref (-1)
 
-let refmt = ref None
-
 
 let js_stdout = ref true
