+++ jscomp/common/js_config.mli
@@ -87,8 +87,6 @@ val force_cmi : bool ref
 val force_cmj : bool ref
 
 val jsx_version : int ref
-val refmt : string option ref
-
 
 val js_stdout : bool ref
