+++ jscomp/snapshot.ninja
@@ -19,7 +19,7 @@ o snapshot: phony  $SNAP/whole_compiler.ml   $SNAP/bsb_helper.ml $SNAP/bsb.ml $S
 # $SNAP/bspp.ml
 o $SNAP/whole_compiler.ml: bspack | ./bin/bspack.exe $LTO
     flags = ${releaseMode} -D BS_ONLY=true -bs-MD -module-alias Config=Config_whole_compiler -bs-exclude-I config  -I $OCAML_SRC_UTILS -I $OCAML_SRC_PARSING -I $OCAML_SRC_TYPING -I $OCAML_SRC_BYTECOMP -I $OCAML_SRC_DRIVER   $includes
-    main = Js_main
+    main = Rescript_compiler_main
     post_process = && node $LTO $SNAP/whole_compiler.ml
 
 o $SNAP/bsb_helper.ml: bspack | ./bin/bspack.exe $LTO
