+++ jscomp/test/test2.js
@@ -1,0 +1,11 @@
+'use strict';
+
+
+var H = {};
+
+var U = {
+  H: H
+};
+
+exports.U = U;
+/* No side effect */
