+++ jscomp/test/test2.ml
@@ -1,0 +1,5 @@
+module U = struct 
+  module H = struct   
+    type t = string  
+  end 
+end
