+++ lib/4.06.1/unstable/bspack.ml
@@ -4011,7 +4011,7 @@ let message = function
   | Bs_derive_warning s ->
       "bs.deriving warning: " ^ s 
   | Bs_fragile_external s ->     
-      s ^ " : the external name is inferred from val name is unsafe from refactoring when changing value name"
+    s ^ " : using an empty string as a shorthand to infer the external's name from the value's name is dangerous when refactoring, and therefore deprecated"
   | Bs_unimplemented_primitive s -> 
       "Unimplemented primitive used:" ^ s
   | Bs_integer_literal_overflow -> 
@@ -11959,8 +11959,6 @@ val force_cmi : bool ref
 val force_cmj : bool ref
 
 val jsx_version : int ref
-val refmt : string option ref
-
 
 val js_stdout : bool ref 
 
@@ -12050,8 +12048,6 @@ let force_cmi = ref false
 let force_cmj = ref false
 
 let jsx_version = ref (-1)
-
-let refmt = ref None
 
 
 let js_stdout = ref true
